Remove generated Typography documentation.
Summary:
It's unlikely that we'll check in the generated docs to the repo directly, so I'm removing the Typography/docs folder.

We should be able to generate the docs at will and to be able to look at them on our local machine. I'm working on this for T371.

Reviewers: #mux, randallli

Reviewed By: #mux, randallli

Subscribers: ajsecord

Projects: #mux, #material_typography

Differential Revision: http://mux-oss-phabricator.appspot.com/D18