Add gendocs script.
Summary:
This script will execute jazzy on any directory within the mdc repo that contains a .jazzy.yaml file. The output of the documentation will live in a docs/ folder.

It also generates an index.json file which could be used by a documentation generator to build the top-level table of contents.

Reviewers: #mux, ajsecord

Reviewed By: #mux, ajsecord

Subscribers: randallli, iangordon

Projects: #mux, #material_components_ios

Maniphest Tasks: T371

Differential Revision: http://mux-oss-phabricator.appspot.com/D24