Catalog by convention.
Summary:
Typically speaking, a structured catalog app will show a list of components and allow the user to drill down to a list of examples for each component. Such an app might be built with a UITableView or UICollectionView, such that the following process is followed to add a new example:

1. Create the new view controller.
2. Add a new entry in the respective component's "examples" list.
3. Upon tapping the entry, create an instance of the view controller and push/present it.

## Convention over configuration

In the interest of encouraging the creation of more examples, I'd like to pursue a solution that *minimizes the number of steps required to create new examples*. The ideal outcome of this task: adding a new example to our 'catalog' is a single step:

1. Create a view controller driving your demo.

Notable constraints to satisfy:

- Zero code written outside of the context of the demo view controller by the demo writer. Someone adding a new example should not have to make any changes to the Catalog's structure.

This diff solves the above with approximately 150 lines of code.

## What's going on in this change?

Deleting the Ink example project. In this new "by convention" approach we simply have to add a ViewController.h/m into the component's `examples/` directory. When `pod install` is ran on the catalog, CocoaPods will detect these files and add them to the build.

Introducing the new `catalog/` directory. This is where the code for the new convention-based catalog app lives.

The MDCCatalog app includes an AppDelegate and a NodeViewController. The AppDelegate enumerates the runtime for classes that respond to the `catalogHierarchy` method and then creates a UINavigationController pointing at the root node of the catalog hierarchy. NodeViewController simply represents the children of a node and allows the user to tap a child row to move deeper in the tree.

There are a few methods in the "Runtime" for both Objective-C and Swift. These are the workhorses for enumerating the classes at runtime.

Reviewers: #material_components_ios_owners, randallli

Reviewed By: #material_components_ios_owners, randallli

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D119