Add CHANGELOG.md.
Summary:
This file tracks changes over time. It should be considered a client-facing
living document explaining changes over time on a release-by-release basis.

This file will contain not only the summarized changes that occur,
but also any migration plans that are necessary from one version to another.

The expectation is that this file will be updated with nearly every change
landing in the `develop` branch.

Reviewers: #material_components_ios_owners, randallli, ajsecord

Reviewed By: #material_components_ios_owners, randallli, ajsecord

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D122