[FlexibleHeader] Always set intensityDidChangeBlock when changing shadow layer.
Summary:
Prior to this change it was possible to change the intensity block and then change
the layer (without providing a new intensity block). The flexible header would then
incorrectly use the old block.

This change ensures that the code path for setting the shadow layer always sets the
intensity block.

Reviewers: junius, #mdc_ios_owners

Reviewed By: junius, #mdc_ios_owners

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D148