Add generate_jazzy_yamls script.
Summary:
Having a .jazzy.yaml in each component's directory has some technical
advantages. Keeping these files up-to-date, however, is not trivial.

This diff includes a script for regenerating all of the .jazzy.yaml
files. The intent is for this script to be run in the following
situations:

- New component added.
- Jazzy version change.

Reviewers: #mdc_ios_owners, ajsecord

Reviewed By: #mdc_ios_owners, ajsecord

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D255