[FlexibleHeader] Poke the header into laying out its content when the view controller has been fully registered.
Summary:
This resolves odd behavior that is apparent if you set the tracking scroll view before setting up forwarding of UIScrollViewDelegate events to the header view.

When the tracking scroll view is set we modify the edge insets. This usually results in a scroll view didScroll event being fired. If the flexible header does not receive this event due to event forwarding not being set up, then the flexible header can be presented in an inconsistent state.

This diff ensures that the flexible header always syncs its state to the tracking scroll view before being presented.

Reviewers: #mdc_ios_owners, ajsecord

Reviewed By: #mdc_ios_owners, ajsecord

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D330