[FlexibleHeader] Always project the flexible header's frame onto the tracking scroll view.
Summary:
This handles the case of the flexible header being a subview of its tracking scroll view. By using convertFrame we're able to "project" the flexible header onto the tracking scroll view.

Before this change we were always assuming that the flexible header's frame was distinct from the tracking scroll view. This works fine when the flexible header is a sibling to the tracking scroll view. It breaks when the flexible header is a subview of the tracking scroll view (e.g. UITableViewController).

Closes https://github.com/google/material-components-ios/issues/99.

Reviewers: #mdc_ios_owners, ajsecord

Reviewed By: #mdc_ios_owners, ajsecord

Subscribers: ajsecord

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D345