[FlexibleHeader] Remove the internal contentOffset property.
Summary: The intent of this property was to mirror the value of the trackingScrollView's contentOffset. It is sufficient to read the value directly from the trackingScrollView. Doing so also resolves a class of errors caused by _contentOffset becoming out of sync with the trackingScrollView's actual contentOffset.

Reviewers: #mdc_ios_owners, ajsecord

Reviewed By: #mdc_ios_owners, ajsecord

Projects: #material_components_ios

Differential Revision: http://codereview.cc/D769