Add support for Cocoapods 1.0.0.
Summary:
DO NOT LAND until after we have finalized our May 25 2016 release.

Cocoapods 1.0.0 was used to generate Pod* and xcode* files.

These changes are NOT compatible with 0.39.0.
These changes may NOT be compatible with version of Xcode 7.2 and earlier.

pod lib lint generates a list of issues that should be addressed.

--OLD INFO--

Cocoapods 0.39 was used to generate the Podfile.locks.

MaterialComponents - Build properly
MaterialUnitTests - Test properly
MaterialCatalog - Fails to Build

The errors (cut after the first 20) are similar to:
<module-includes>:1:9: note: in file included from <module-includes>:1:
#import "Headers/MaterialComponentsCatalog-umbrella.h"
        ^
/Users/iangordon/projects/material-components-ios/catalog/Pods/Target Support Files/MaterialComponentsCatalog/MaterialComponentsCatalog-umbrella.h:10:9: note: in file included from /Users/iangordon/projects/material-components-ios/catalog/Pods/Target Support Files/MaterialComponentsCatalog/MaterialComponentsCatalog-umbrella.h:10:
#import "CollectionCellsLayoutExample.h"
....

Reviewers: featherless, randallli, ajsecord, #material_components_ios

Reviewed By: featherless, randallli, ajsecord, #material_components_ios

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D873