[Slider] Fix visual bug on iPhone 6+ disabled state: hairline behind thumb should not be there.
Summary:
Fixes https://github.com/google/material-components-ios/issues/566 by extending the mask for a MDCThumbTrack by 1 pixel on the top and bottom, to avoid a half pixel showing through the mask. To stop this change from affecting the way MDCSwitch renders, I altered the rounded corners of the thumb track to be rendered via the cornerRadius property of the track view, instead of using the mask, as it was previously.

Reason for this bug is the behavior of CAShapeLayer, which is used to mask off the necessary parts of the track view. According to [Apple's documentation](https://developer.apple.com/library/ios/documentation/GraphicsImaging/Reference/CAShapeLayer_class/), "Shape rasterization may favor speed over accuracy. For example, pixels with multiple intersecting path segments may not give exact results."

This has the result that when the track view falls on a subpixel, the track view does a better job rasterizing and handling this situation than its CAShapeLayer mask, which results in half-shaded pixels peaking out from underneath the mask. This appeared specifically in iPhone 6+ because of the way that device scales points to pixels, resulting in subpixels where on other devices there were none.

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D955