[ThumbTrack] Make slider and switch more responsive by not forcing a slop distance before movements are recognized
Summary:
Makes the Slider and Switch components more responsive by removing the behavior of waiting a set "slop" distance before recognizing the movement of the thumb. This was due to our use of the UIPanGestureRecognizer, which enforces the minimum movement distance internally.

Turns out it's a really easy switch to swap in a UILongGestureRecognizer instead, which recognizes movements instantly and includes the exact same updates as UIPanGestureRecognizer, so I did that.

Solution found here: http://stackoverflow.com/questions/10728066/why-is-there-a-delay-when-moving-object-using-uipangesturerecognizer

Closes #553.

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D964