[Slider] Make thumb view hollow when on the minimum value.
Summary:
Makes thumb view hollow when the thumb is on the minimum value. Animates into this state when you move the thumb to the minimum value.

{F5135}

This update includes a sizable refactor of how MDCThumbTracks are updated. I split up the three phases of updating the views into:

1. No-animation block
2. Main animation block
3. Secondary animation block

The first applies changes that you want to see reflected instantly, the second does the main animation of moving the thumb and "on" layer, and the third handles the secondary animation of the thumb animating into this hollow minimum value state. This animation is more generally defined as the animations on the thumb that occur after the thumb reaches its position on the track.

Callers can now specify when they update the thumb whether they want the thumb move to be animated, and separately whether the thumb should animate after arriving at its destination. This way, we can handle the difference between taps on the track, drags of the thumb, and code-based non-animation updates. With taps we want both to be animated, with dragging we want only secondary update to be animated, and if you call `-updateValue:` directly, neither should be animated.

This refactor accounts for those cases, and makes the code a little easier to understand in my opinion.

Closes #555

Relies on D956

Reviewers: ajsecord, O1 Material components iOS, randallli

Reviewed By: ajsecord, O1 Material components iOS, randallli

Subscribers: ajsecord

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D963