[Slider] Grow thumb when dragging the slider
Summary:
Grows the thumb view when a user is dragging it. Animates into the larger state and back into the regular size after letting go of the thumb.

{F5288}

To support this, I also added an attribute `thumbGrowsWhenDragging` to `MDCThumbTrack` so we can enable this feature for `MDCSlider` while leaving `MDCSwitch` unaffected.

Closes #556

Relies on D963 and D989

Reviewers: randallli, O1 Material components iOS

Reviewed By: randallli, O1 Material components iOS

Subscribers: randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1018