[Slider] Refactored how the slider animates across an anchor value
Summary:
Currently `MDCThumbTrack` has a property `filledTrackAnchorValue` describing the value from which the "on track" should be drawn. This is not currently used in the Slider or the Switch, and it's not exposed by any public API. However it could potentially it would be useful down the line for the Slider.

With this in mind, the thumb track currently handles a special case when animating the position of the thumb: when the previous and next value span the anchor point. It separates the animation into two, first animating to the anchor point, and then to the next value. The problem, as @randallli pointed out in the review of D963, is that both animations ease in and out, so you end up with a strange stuttering effect.

I fixed that, making the first animation only ease in, and the second ease out, resulting in an overall effect of the ease-in-out expected in this animation. The calculation of the duration of each animation was also flawed, so I adjusted it. I also fixed the rendering of the anchor point in general, which previously varied by a few pixels, resulting in a misleading view of the slider.

Not sure of the exact process for changing an existing public API, but given that this now actually works, it would be interesting to expose the property on `MDCSlider`.

An open question is where to make the thumb be hollow in this setup. I haven't changed anything, so it currently behaves like this:

{F5296}

{F5297}

Common sense would dictate that it should be the second state that has the hollow circle appearance, but I'm not a designer! Any idea who to talk to about that?

Relies on D963

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Subscribers: randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1022