[Switch] Fixed ink not displaying on Switch
Summary:
Due to an earlier change I made for this component, the ink that should have been displayed on MDCSwitch was not displaying. This fixes that by switching to UIResponder events on MDCSwitch instead of the UIControl events that I was using previously.

A quick summary of the issue is that because we were using those UIControl events, `userInteractionEnabled` had to be set to `NO` for the thumb view, which handles ink displaying via a gesture controller. The only way to turn that back to `YES` was to switch to using UIResponder events, which allow a view to handle touches as well as pass those same touches through to subviews.

To go along with this, I also made our touch handling more robust, to ensure behavioral parity with the slider on the previous UIControl events. This means tracking touches' positions to make sure we handle only one touch event at a time, and ignore other events when needed.

Before (when touching):

{F5398}

After:

{F5399}

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Subscribers: randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1050