[Switch] Send ValueChanged events when dragging, not just tapping
Summary:
Before the Switch would only send events about value changed in response to taps. This didn't really make much sense and resulted in interactions like this:

{F5465, media=video}

When they should have been like this:

{F5464, media=video}

So now it sends the event for dragging, when the touch is ended or cancelled.

Reviewers: randallli, O1 Material components iOS

Reviewed By: randallli, O1 Material components iOS

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1077