[Snackbar] The button a11y identifier is the action's
Summary:
Prior to this Diff, the accessibility identifier for the button was
incorrectly set to the accessibility identifier of the whole message,
while it needs to be the accessibility identifier of the action object
only.

Closes https://github.com/google/material-components-ios/issues/622.

Test Plan:
Check that the accessibility identifier of the button is the same as
the message action's.

Reviewers: junius, ajsecord, O1 Material components iOS

Reviewed By: ajsecord, O1 Material components iOS

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1134