[Snackbar] Align text correctly in RTL layout
Summary:
Thanks to Autolayout, most of the work is done. This Diff only sets the
text alignment to Natural for the text label of the snackbar.

Test Plan:
Open Catalog in an RTL language (or pseudo RTL language setting in Xcode) and
check that snackbars, and especially snackbars with buttons are laid out from
right to left.

Reviewers: junius, O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1168