[Slider] Expanded slider demo within a collection view.
Summary:
Switched the default demo for MDCSlider to a new view controller that displays several different types of sliders in a collection view format. Serves as an example of how to use a slider in this context, as collection views are often used for settings screens where sliders are most likely to appear.

Also shows off new options for the slider, such as displaying the discrete value label or not, and setting an anchor value to draw the track from.

{F6385}

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1486