[Slider] Property for whether or not thumb is hollow at minimum value
Summary:
Adds `thumbIsHollowAtStart` property to MDCSlider, to support usages where you don't need the thumb view to be hollow when `slider.value == slider.minimumValue`, i.e. when the slider is used as a scrubber in an audio or video file.

Also adds an example of this property being `NO` to the primary slider example view controller.

{F6417}

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1518