Small MDCShadowLayer fixes
Summary:
`renderInContext:`
This was broken for a number of reasons. This method generally doesn't need to be overriden, and calling it once will call renderInContext on all child layers. If you need to lay out a layer before using `renderInContext` you should also call `layoutIfNeeded`. Setting the `backgroundColor` on the MDCShadowLayers unfortunately did not cause them to render and would be incorrect anyway.

Reviewers: O1 Material components iOS, randallli

Reviewed By: O1 Material components iOS, randallli

Tags: #material_components_ios

Differential Revision: http://codereview.cc/D1526