Fixes issue where the rotation animation was broken for any GOOHeaderViewController child’s view.
The cause was that -layoutIfNeeded was being called on a number of sublayers of the flexible header in its own -layoutSubviews method. This was causing the child view controller’s view to get laid out immediately before the UIKit internal rotation-snapshotting mechanism is activated, causing the animation to be uncentered and incorrect. The fix is to just commit the CATransaction correctly to update the layers. There is no need to call -layoutIfNeeded on them.

PiperOrigin-RevId: 156618621