Add Masked Transition component. (#1513)
* Add Masked Transition component.

This component makes it possible to present a view controller from a source view, such as a floating action button, using a masked reveal transition.

This component depends on Material Motion's Transitioning, MotionInterchange, and MotionAnimator libraries.

The component itself is a Transition instance and can be used like this:

    vc.transitionController.transition = MDCMaskedTransition(sourceView: fab)
    present(vc, animated: true)

The component's motion is defined in its motion spec and was extracted directly from the Material motion spec.

The transition supports the following contexts:

- Fullscreen
- Bottom sheet
- Floating card
- Bottom toolbar

* Code review feedback.

* Add README.md.

* Copy the frame calculation block.

* Reword the overview to reduce implementation details.

* Remove scrim sharing and document full screen behavior.

* Audit rect math and remove example debug logs.

* Update to use hypot.

* Remove presentation controller storage.

* Update podspec with new MotionTransitioning dependency.

* Fix autoresizing bugs and update header import names.

* Fix floating point conversion warning.

* Use bottomCard spec for all positions.