[Fonts] Cache system fonts (#1689)
`[UIFont systemFontForSize:weight]` performs some kind of internal
caching (it returns the same pointers) but the performance is relatively
poor compared to an actual NSCache. Introducing a cache in the
MDCSystemFontLoader and clearing the cache whenever the Content Category
size changes (dynamic type notifications) even though we're not
currently using those fonts.

Performance is improved when scrolling MDCCollectionViewTextCells, which
retrieve the fonts in `prepareForReuse`.