[TextFields] Multiline textfields (#1756)
* [TextFields] Multiline text fields element (#1737)

* [TextFields] Multiline text field element.

* [TextFields] Formatting,

* [TextFields] Adding missing change for multiline protocol.

* [TextFields] Multiline added to main header. (#1740)

* [TextFields] Multiline tests. (#1741)

* [TextFields] Controllers. (#1742)

* [TextFields] Controllers.

* [TextFields] Multiline clear button correction.

* [TextFields] Alphabetizing.

* [TextFields] Multiline examples. (#1743)

* [Scripts] Only lint catalog, components, and demos (#1744)

The lint_all script should stick to the MDC-iOS code and not the docs,
scripts, etc. that are included in the repository.  Specifically, we are
only linting the Components, Catalog, and Demos that we ship.

Closes #1578

* [Collections] Making Disclosure indicator tinted (#1745)

Collection cell accessory views set to "Disclosure Indicator" will now
use template mode and will utilize the tinting of the cell/collection
view.

Closes #1575

* [NavigationBar] Improve centering when barButtonItems are asymmetric (#1721)

* [NavigationBar] Improve centering when barButtonItems are asymmetric

Adjusting the NavigationBar layout computation to determine if there is
sufficient space in the center of the navigation bar before adjusting
the Title frame. Adding NavBar example for varying item sizes.

Closes #1662

* Adding a few comments to describe how we're centering

* [Math] Align frames to pixel boundaries (#1730)

Adding a new function to realign a CGRect (frame) to pixel boundaries
based on the screen scale. After a review of the catalog, I'm updating
the areas where I found misaligned frames.

Closes #1723

* [TextFields] Multiline examples.

* [TextFields] Minor corrections to multiline examples.

* [TextFields] Multiline full width example scrolls when keyboard comes up.

* [TextFields] Removing extra file.

* [TextField] Formatting.

* [TextField] Formatting.

* [TextFields] Formatting. (#1748)

* [TextField] Correcting copyright.

* [TextField] Corrections to fix warnings and errors in multiline example.

* [Text fields] Trailing view support (#1747)

* [TextFields] Trailing view work.

* [TextFields] Multiline trailing view layout.

* [TextFields] Formatting.

* [TextFields] Spelling correction.

* [TextFields] Comment.