[Snackbar] Fixing snackbar/overlay during rotation (#1830)
The MDCSnackbarOverlayView was not correctly laying-out its
containingView during rotation. Forcing the containingView to layout
during the overlay view's layoutSubviews method has corrected the
position of the Snackbar for notification recipients. It was also using
the fixedCoordinateSpace of UIScreen, which is incorrect as it does not
account for rotation.

Updating the Snackbar Overlay View example to include an animated FAB to
demonstrate an incompatibility with autoresizing masks.

Closes #1701