[Buttons] Default hitAreaInsets for FAB (#1866)
The Floating Action Button should have default hitAreaInsets so that it
adheres the minimum touch target in the Material Design guidelines.
This commit introduces a set of default hitAreaInsets for the different
FAB styles as well as a warning if a button's touch target is smaller
than the 48-point minimum.

Closes #223