[Dialogs] Migrate to Material Motion. (#2481)
This change implements the Material Motion spec for the Dialogs component using the Motion Interchange, Animator, and Transitioning APIs.

Many of the existing APIs will be able to be deprecated or made private in the future, though there is considerable existing downstream client usage of many of these APIs so this will need to be done with some care.

The client migration path for this change will look something like this on average:

```objc
// Old API
_transitionController = [[MDCDialogTransitionController alloc] init];
self.transitioningDelegate = _transitionController;		
self.modalPresentationStyle = UIModalPresentationCustom;

// New API
self.mdm_transitionController.transition = [[MDCDialogTransition alloc] init];
```