[NavBar, ButtonBar] Add NavBar tests, fix ButtonBar KVO (#2713)
* [NavBar, ButtonBar] Add NavBar tests, fix ButtonBar KVO

ButtonBar was not correctly setting up KVO when it was created with an NSCoder (restored from an
archive). In that case, dealloc would generate an exception and crash.  Instead of writing to the
backing iVar during initWithCoder:, the ButtonBar should call through to its setter instead.  This
is less safe (because the subclass may not be initialized) but guarantees observing the objects it
is restoring.

Added a simple NavBar encoding test.