[BottomNavigation] Fix memory leak in example (#3017)
The `BottomNavigationTypicalUseExample` has a retain cycle in the
`dispatch_after` block it uses to update the badge value. This causes
example VCs to be accumulated if they are opened repeatedly. A simple
fix is to keep a weak reference in the dispatched block making the next
update call a no-op.