[ButtonBar] Remove UIAppearance call in themer (#3020)
MDCButtonBarColorThemer was explicitly calling UIAppearance for MDCButtonBar when contained in an MDCNavigationBar to clear the backgroundColor property. This is "magic" behavior and should not be included in the themer. Instead, the navigation bar will force the buttonBar background colors to nil (clear). Since having ButtonBars is an implementation detail, clients should not expect them to have a different background color than the navigation bar.

Closes #2789