[Dialogs] Fixed regression where in some cases tapping on the buttons in the alert would not do anything (#3182)
* Fixed MDCPageControl crasher when resetting -numberOfPages to 0. This is supported by UIPageControl.

Crasher details:
Test Case '-[PageControlExampleTests testResetNumberOfPagesToZero]' started.
~/material-components-ios/components/PageControl/tests/unit/PageControlExampleTests.m:139: error: -[PageControlExampleTests testResetNumberOfPagesToZero] : failed: caught "NSRangeException", "*** -[__NSArrayM objectAtIndexedSubscript:]: index 0 beyond bounds for empty array"
(
	0   CoreFoundation                      0x000000010e3ed03b __exceptionPreprocess + 171
	1   libobjc.A.dylib                     0x000000010da81f41 objc_exception_throw + 48
	2   CoreFoundation                      0x000000010e42ce2c _CFThrowFormattedException + 194
	3   CoreFoundation                      0x000000010e41e714 -[__NSArrayM objectAtIndexedSubscript:] + 148
	4   MaterialComponents                  0x0000000128271441 -[MDCPageControl resetControl] + 2193
	5   MaterialComponents                  0x000000012826e489 -[MDCPageControl setNumberOfPages:] + 281

* Also fixed crasher when the client sets a value for -currentPage (e.g. zero) when -numberOfPages is zero.

* Improved the unit test to hit the 2nd crasher prior to the fix

* Cleaner unit test for the crashers per review feedback

* [Slider] Fixed the numerical value label continually increasing in size when sliding the thumb

* Fixed typos in comment

* Fixed typos in comment (take 2)

* Fixed typos in comment (take 2a)

* Slightly more elegant code

* [AlertController] Fixed regression where in some cases tapping on the buttons in the alert would not do anything

Specifically, we are seeing this behavior (b/77155189 and b/77152040) when a text field is first responder when the alert is presented. When that happens, -[MDCAlertController actionButtonPressed:] (ie. the method expected to be called by UIKit) doesn't get called when the user taps on the buttons.

Per UIKit's documentation, specifying a nil target in -[UIControl addTarget:action:forControlEvents:] is supported ("passing in nil as the target goes up the responder chain"), however it seems to me that it is preferable to be explicit in order to be more robust to the responder chain being not exactly the way it should be.

This is also the previous behavior of MDCAlertController prior to https://github.com/material-components/material-components-ios/commit/b22c8c2650a4a5b21c76cf37b3d1041de639e4e5