[Catalog] Remove the global theme change notifications. (#3213)
We may add this back in a more well-defined manner in the future, but for now we're removing this pattern.

For future reference, this pattern was implemented in order to allow examples to change the global app theme because examples live outside of the app theme's target space.

In the future, we may add a theme changer to the catalog target so that we don't have to do odd cross-target juggling.

Pivotal story: https://www.pivotaltracker.com/story/show/156449674