[kokoro] Disable verbose output on bazel builds. (#3262)
The non-verbose output should typically be sufficient to identify any test or build failures. We had originally been enabling verbose output because some of the failures were happening further upstream in the build toolchain, but now that kokoro has largely stabilized we can step down the verbosity.

If any individual PR needs increased verbosity to identify a failure then `VERBOSE_OUTPUT` can be manually enabled in the script as part of the PR.