[ThumbTrack] Add support for colored tick marks (#3329)
The Slider needs a way to color the tick marks within the "filled"
region differently from the tick marks in the "unfilled" region of the
track. The ThumbTrack should support an API that allows customization of
the colors and definition of the "filled" or "active" track area.

**Resulting Effect (with non-default colors)**
![slider-colored-dots](https://user-images.githubusercontent.com/1753199/38683860-63175184-3e3c-11e8-9014-7c55a528a942.gif)

**Before**
![slider-ticks-before](https://user-images.githubusercontent.com/1753199/38697970-796153de-3e61-11e8-9c7e-6dc29755e381.png)

**After**
![slider-ticks-after](https://user-images.githubusercontent.com/1753199/38697976-7cbf6084-3e61-11e8-9cf5-e873abffdef4.png)



Partially implements #3137
Pivotal story: https://www.pivotaltracker.com/story/show/155525171