[NavigationBar] Fix bug where system font traits would be lost. (#3360)
### Context

When setting a font on MDCNavigationBar, we force the font to have a size of 20.

### Prior to this change

We were enforcing the font size by using an older UIFont API - fontWithName:size:. This API does not preserve font descriptor attributes from the provided font, it only preserves the font name and size.

This was resulting in the loss of an important system font descriptor attribute, namely `NSCTFontUIFontDesignTrait` which was set to `NSCTFontUIFontDesignDefault`. This font attribute affects the kerning of fonts when displayed on screen.

### After this change

We are now using the iOS 7 API - fontWithDescriptor:size:. This API preserves all of the underlying font descriptor attributes.

The result of this change is an intentional visual change.

As part of this behavioral change, I've also updated the related unit tests to enforce this behavior in the future.

To see a visual depiction of the change in behavior, open the following two screenshots and flip between them:

Before this fix:
![lossy](https://user-images.githubusercontent.com/45670/38755649-8a172a22-3f34-11e8-9f53-15ca6789e351.png)

After this fix:
![notlossy](https://user-images.githubusercontent.com/45670/38755654-8d998fa0-3f34-11e8-8112-49e3d4b7d76e.png)

Visual delta:
![difference](https://user-images.githubusercontent.com/45670/38755656-91d02f5c-3f34-11e8-8032-36d3ffc8f52e.png)