[kokoro] Add website generator command (#3541)
This will allow our PRs to generate the documentation website as part of our presubmits. This will let us catch any errors that might otherwise affect the ability for the website to be deployed.

This PR also updates the site generator's git url to an https one so that the kokoro instance can clone the repo.