[Chips] Vertically center chips within available space (#3511)
When a Chip Field's `chipHeight` value is greater than the actual height of a
chip, the chip is not aligned with the input text. Although baseline is ideal
(to ensure multiple fonts work, etc.), vertically centering works sufficiently
well when the text field and the chip have the same font.

Partially implements #3507

**Before**
![chipfield-size-before](https://user-images.githubusercontent.com/1753199/39321042-d532df64-4953-11e8-85a3-e7108a1106a6.png)

**After**
![chipfield-size-after](https://user-images.githubusercontent.com/1753199/39321048-d8e33e06-4953-11e8-9845-4fb38964886b.png)