Remove redundant copy from our root readme. (#3569)
This content is prone to getting stale and is largely a copy of what's found in docs/docsite-index.md. Instead of copying it into the README.md we now link to the documentation website. We may want to consider increasing the visibility of the documentation link as well.

Pivotal story: https://www.pivotaltracker.com/story/show/156775556