[Lists] Added initial example for implementing the List component (#3646)
Pivotal: https://www.pivotaltracker.com/story/show/155597168

This example uses auto layout constraints to show a way to use UICollectionViewCells with a vanilla UICollectionViewController to achieve a List Cell that is self-sizing, supports dynamic type, RTL, and orientation changes.

This example also supports the correct layout with and without an image set in the cell, and also with an unset titleLabel.text and an unset detailsLabel.text.