[Catalog] Finish menu functionality and placement (#3747)
Pivotals: https://www.pivotaltracker.com/story/show/156998143 (settings) - closes story
https://www.pivotaltracker.com/story/show/157165663 (menu accessible from main screen) - closes story
https://www.pivotaltracker.com/story/show/157022469 (red lines)

In this PR these are the changes:

Added menu button to main screen
Updated the main screens flexible header redlines after adding the button so the layout will look good
Removed the versioning from the title in the main screen as per designs
Added a setting menu that was previously shown by tapping the app 3 times
Remove the 3 tap trigger for the settings menu
Removed the "do not show this again" in the settings menu as now we can't trigger it mistakenly with a triple tap.
Added the version number instead of where the help was in the menu dialog so we can remove it from the main screen title and be able to also access it anywhere in the app.