[kokoro] Add danger support. (#3757)
Adds support for running Danger checks against PRs.

## What is Danger?

[Danger](http://danger.systems/js/) is a tool built and maintained by @orta at Artsy. From the site docs:

> Danger runs during your CI process, and gives teams the chance to automate common code review chores.
> 
> This provides another logical step in your build, through this Danger can help lint your rote tasks in daily code review.
> 
> You can use Danger to codify your teams norms. Leaving humans to think about harder problems.
> 
> This happens by Danger leaving messages inside your PRs based on rules that you create with JavaScript or TypeScript.
> 
> Over time, as rules are adhered to, the message is amended to reflect the current state of the code review.

## Example output

<img width="804" alt="screen shot 2018-05-09 at 11 50 34 am" src="https://user-images.githubusercontent.com/45670/39825186-35a31a64-537f-11e8-96ae-dc4a520cf28e.png">

## How does it work?

Danger runs as part of our CI workflow. Danger has a dedicated job that runs in parallel with our other kokoro jobs. When the kokoro job runs, it invokes our `.kokoro` script with `danger` as the environment. The danger "logic" is found in Dangerfile.js. To start, our dangerfile simply checks whether a PR has a description that is longer than 10 characters. Over time, I hope that we can expand to more checks such as:

- Recommending that source changes be matched by test changes.
- Reminding to regenerate the component README.md if any of the doc articles change.
- Reminding that podspec and BUILD file changes often have to go hand-in-hand.
- Enforcing use of umbrella header imports over direct header imports.

## Can I run the danger checks locally?

Danger runs on PRs, so you won't be able to preview warnings before pushing your changes to GitHub. You technically can run the danger CI locally, but it will likely do the same work as the kokoro job (which starts running once you push the PR).

## What happens if danger gives a warning and I don't have any new changes to push?

For now, Danger is simply a warning, so it won't block any PRs from landing. But if you want to make sure that your PR is clean you can restart the job by clicking the Danger status bit in the PR and restarting the job for your PR.