[FlexibleHeader] Always update opacity for views that hide when shifted. (#4170)
* [FlexibleHeader] Always update opacity for views that hide when shifted.

Before this change, views that hide when shifted would only change their visibility if shifting was enabled. This resulted in a bug where if shifting was disabled then views would not become visible again once the header shifted back on screen.

After this change, views that hide when shifted will always change their visibility when the header's frame changes.

* Simplify the logic.

* ORder.