[kokoro] Add new apidiff CI command. (#4245)
This command will generate an API diff for the current pull request. In a follow-up PR it will also post the results of the API diff to the PR as a comment.

This is the first of two changes to the `apidiff` branch that will add support for API diff reporting to our PRs. The second PR will add the github comment reporting tool.

This was previously reviewed in https://github.com/material-components/material-components-ios/pull/4181.